FE-1642: Make the entities tree the only left panel rendering
