+++ .changeset/entities-tree-left-panel.md
@@ -1,0 +1,5 @@
+---
+"@hashintel/petrinaut": patch
+---
+
+The left panel shows the Entities tree for everyone. The stack of separate Nodes, Token Types, Differential Equations and Parameters lists is gone, along with the "Entities tree view" setting that chose between them. Nets with subnets keep the Subnets panel below the tree.
