+++ libs/@hashintel/petrinaut/src/ui/constants/ui-subviews.ts
@@ -11,24 +11,12 @@ import { diagnosticsSubView } from "../views/Editor/panels/BottomPanel/subviews/
 import { simulationSettingsSubView } from "../views/Editor/panels/BottomPanel/subviews/simulation-settings";
 import { actualTimelineSubView } from "../views/Editor/panels/BottomPanel/subviews/simulation-timeline/actual";
 import { simulationTimelineSubView } from "../views/Editor/panels/BottomPanel/subviews/simulation-timeline/main";
-import { differentialEquationsListSubView } from "../views/Editor/panels/LeftSideBar/subviews/differential-equations-list";
 import { entitiesTreeSubView } from "../views/Editor/panels/LeftSideBar/subviews/entities-tree";
 import { netsListSubView } from "../views/Editor/panels/LeftSideBar/subviews/nets-list";
-import { nodesListSubView } from "../views/Editor/panels/LeftSideBar/subviews/nodes-list";
-import { parametersListSubView } from "../views/Editor/panels/LeftSideBar/subviews/parameters-list";
-import { typesListSubView } from "../views/Editor/panels/LeftSideBar/subviews/types-list";
 
 import type { SubView } from "../components/sub-view/types";
 
 export const LEFT_SIDEBAR_SUBVIEWS = [
-  nodesListSubView,
-  typesListSubView,
-  differentialEquationsListSubView,
-  parametersListSubView,
-  netsListSubView,
-] as const satisfies SubView[];
-
-export const LEFT_SIDEBAR_TREE_SUBVIEWS = [
   entitiesTreeSubView,
   netsListSubView,
 ] as const satisfies SubView[];
