--- libs/@hashintel/petrinaut/src/ui/views/Editor/panels/LeftSideBar/subviews/types-list.tsx
@@ -1,169 +1,0 @@
-import { use } from "react";
-
-import { Button } from "@hashintel/ds-components";
-
-import { usePetrinautMutations } from "../../../../../../react/hooks/use-petrinaut-mutations";
-import { ActiveNetContext } from "../../../../../../react/state/active-net-context";
-import { EditorContext } from "../../../../../../react/state/editor-context";
-import { SDCPNContext } from "../../../../../../react/state/sdcpn-context";
-import { useIsReadOnly } from "../../../../../../react/state/use-is-read-only";
-import { TokenTypeIcon } from "../../../../../constants/entity-icons";
-import { UI_MESSAGES } from "../../../../../constants/ui-messages";
-import {
-  RowMenu,
-  createFilterableListSubView,
-} from "./filterable-list-sub-view";
-
-import type { SubView } from "../../../../../components/sub-view/types";
-
-// Pool of 10 well-differentiated colors for types
-const TYPE_COLOR_POOL = [
-  "#3b82f6", // Blue
-  "#ef4444", // Red
-  "#10b981", // Green
-  "#f59e0b", // Amber
-  "#8b5cf6", // Violet
-  "#ec4899", // Pink
-  "#14b8a6", // Teal
-  "#f97316", // Orange
-  "#6366f1", // Indigo
-  "#84cc16", // Lime
-];
-
-/**
- * Get the next available color from the pool that's not currently in use.
- * If all colors are in use, cycle back to the beginning.
- */
-function getNextAvailableColor(existingColors: string[]): string {
-  const unusedColor = TYPE_COLOR_POOL.find(
-    (color) => !existingColors.includes(color),
-  );
-  return unusedColor ?? TYPE_COLOR_POOL[0]!;
-}
-
-/**
- * Extract the highest type number from existing type names.
- * Looks for patterns like "Type 1", "Type 2", "New Type 3", etc.
- */
-function getNextTypeNumber(existingNames: string[]): number {
-  let maxNumber = 0;
-  for (const name of existingNames) {
-    // Match patterns like "Type 1", "New Type 2", etc.
-    const match = name.match(/Type\s+(\d+)/i);
-    if (match) {
-      const num = Number.parseInt(match[1]!, 10);
-      if (num > maxNumber) {
-        maxNumber = num;
-      }
-    }
-  }
-  return maxNumber + 1;
-}
-
-/**
- * TypesSectionHeaderAction renders the add button for the types section header.
- */
-export const TypesSectionHeaderAction: React.FC = () => {
-  const {
-    activeNet: { types },
-  } = use(ActiveNetContext);
-  const { extensions } = use(SDCPNContext);
-  const { addType } = usePetrinautMutations();
-  const { selectItem } = use(EditorContext);
-
-  const isReadOnly = useIsReadOnly();
-
-  const isDisabled = isReadOnly || !extensions.colors;
-  let tooltip = "Add token type";
-  if (isReadOnly) {
-    tooltip = UI_MESSAGES.READ_ONLY_MODE;
-  } else if (!extensions.colors) {
-    tooltip = UI_MESSAGES.EXTENSION_UNAVAILABLE;
-  }
-
-  return (
-    <Button
-      aria-label="Add token type"
-      size="xs"
-      variant="ghost"
-      disabled={isDisabled}
-      tooltip={tooltip}
-      iconName="plus"
-      onClick={() => {
-        const existingColors = types.map((type) => type.displayColor);
-        const existingNames = types.map((type) => type.name);
-        const nextNumber = getNextTypeNumber(existingNames);
-        const nextColor = getNextAvailableColor(existingColors);
-
-        const id = `type__${Date.now()}`;
-        const newType = {
-          id,
-          name: `Type ${nextNumber}`,
-          iconSlug: "circle",
-          displayColor: nextColor,
-          elements: [
-            {
-              elementId: `element__${Date.now()}`,
-              name: "dimension_1",
-              type: "real" as const,
-            },
-          ],
-        };
-        addType(newType);
-        selectItem({ type: "type", id });
-      }}
-    />
-  );
-};
-
-const TypeRowMenu: React.FC<{ item: { id: string } }> = ({ item }) => {
-  const { removeType } = usePetrinautMutations();
-  const { extensions } = use(SDCPNContext);
-  const isReadOnly = useIsReadOnly();
-  const isDisabled = isReadOnly || !extensions.colors;
-
-  return (
-    <RowMenu
-      items={[
-        {
-          id: "delete",
-          text: "Delete",
-          icon: "trash",
-          tone: "error",
-          disabled: isDisabled,
-          onClick: () => removeType({ typeId: item.id }),
-        },
-      ]}
-    />
-  );
-};
-
-/**
- * SubView definition for Token Types list.
- */
-export const typesListSubView: SubView = createFilterableListSubView({
-  id: "token-types-list",
-  title: "Token Types",
-  tooltip: "Manage data types which can be assigned to tokens in a place.",
-  defaultCollapsed: false,
-  resizable: {
-    defaultHeight: 300,
-    minHeight: 200,
-    maxHeight: 600,
-  },
-  useItems: () => {
-    const {
-      activeNet: { types },
-    } = use(ActiveNetContext);
-    return types.map((type) => ({
-      ...type,
-      icon: TokenTypeIcon,
-      iconColor: type.displayColor,
-    }));
-  },
-  getSelectionItem: (type) => ({ type: "type", id: type.id }),
-  renderItem: (type) => type.name,
-  renderRowMenu: TypeRowMenu,
-  emptyMessage: "No token types yet",
-  renderHeaderAction: () => <TypesSectionHeaderAction />,
-});
