+++ libs/@hashintel/petrinaut/src/ui/views/Editor/panels/SimulateView/experiments/create-experiment-drawer.test.tsx
@@ -149,7 +149,6 @@ const TestProviders = ({
     setShowMinimap: () => {},
     setSnapToGrid: () => {},
     setPartialSelection: () => {},
-    setUseEntitiesTreeView: () => {},
     setEnableNetComponents: () => {},
     setEnableNotebookView: () => {},
     setEnableAdHocScenarios: () => {},
